+++ replay/include/ecs/BasicTypeDefs.h
@@ -239,8 +239,8 @@ namespace rendinst {
   constexpr inline uint32_t riex_max_type() { return (1u << (48 - ri_instance_type_shift)) - 1; }
   constexpr inline uint32_t riex_max_inst() { return ri_instance_inst_mask; }
 
-  inline uint32_t handle_to_ri_type(riex_handle_t h) { return h >> ri_instance_type_shift; }
-  inline uint32_t handle_to_ri_inst(riex_handle_t h) { return h & ri_instance_inst_mask; }
+  inline uint32_t handle_to_ri_type(riex_handle_t h) { return (uint32_t)(h >> ri_instance_type_shift); }
+  inline uint32_t handle_to_ri_inst(riex_handle_t h) { return (uint32_t)(h & ri_instance_inst_mask); }
   inline riex_handle_t make_handle(uint32_t ri_type, uint32_t ri_inst)
   {
     return (uint64_t(ri_type) << ri_instance_type_shift) | ri_inst;
